# storage/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'Evolução local de snipe-it API Documentation'
  description: 'This documentation aims to provide all the information you need to work with the Snipe-IT API.'
  version: 1.0.0
servers:
  -
    url: 'https://snipe-it.test'
...
paths:
...
  '/api/v1/categories/{item_type}/selectlist':
    get:
      summary: Selectlist
      operationId: selectlist
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Unauthorized or unauthenticated.'
                properties:
                  error:
                    type: string
                    example: 'Unauthorized or unauthenticated.'
      tags:
        - Categories
    parameters:
      -
        in: path
        name: item_type
        description: ''
        example: architecto
        required: true
        schema:
          type: string
...